make compile_circuits kwargs only
